Add Validation service and cleanup
